Sending and Receiving Data Over Sockets
In this commit I focused on getting the structure of variable storage complete.  This meant getting all variables stored in their respective structs.  I also added in global variable functionality through a global struct. This struct contains all global
variables, as well as a way to track current states.  At the start of each state this struct is sent to a secondary server (right now it just echos a response) for possible storage.  That way important variables are tracked even in the case of failure
of the local program.  Right now it just works locally, but making it work across machines will be completed after approval of this method.