Added Start Functionality
Added a system that allows the programmer to define the start state of the program using a flag within the program itself.  This allows it to not automatically begin execution at the first defined state, or at a state defined in the command line.  To use this feature, place `#start x` (where x is the state name) after the includes section, but before the global variables.  The program will begin executing at that state.  In the absence of that flag, it will start at the first defined state by default.