Adding in single send capabilities
The program now sends data each time a variable is changed.  At first it would compile a list of variables to update, but now it sends each one after it is updated.  This is because there is no time difference between sending them at the end of the state vs. sending them following their update.