Individual Variable Backup & Starting From Saved State
The individual variable backup is now completely implemented into the language.  This includes the ability to define variables as either `volatile` (they will be backed up after change) or `nobackup` (it will only be backed up if the whole global struct is backed up).

The program will also no look for a backup file at start up.  If it finds one, it will then change the global data structure accordingly so that the program can pick up where it left off.